refactor(S1.2.3): refactor code profile and vehicle form
