+++ frontend/src/app/pages/change-password/change-password.css
@@ -1,3 +1,13 @@
 .cursor-pointer {
   cursor: pointer;
 }
+.input-focus {
+  border-radius: 0.375rem;    
+  transition: border 0.2s ease;
+  box-shadow: none;
+}
+
+.input-focus:focus {
+  border: 2px solid #000; 
+  outline: none;
+}
