+++ frontend/src/app/pages/change-password/change-password.html
@@ -6,12 +6,11 @@
 
         <h2 class="text-center mb-4 fw-semibold">Change Password</h2>
 
-        <!-- Current password -->
         <div class="mb-3 position-relative">
           <input
             #currentPwd
             type="password"
-            class="form-control"
+            class="form-control input-focus"
             placeholder="Current password"
             [type]="showCurrent ? 'text' : 'password'"
           />
@@ -25,12 +24,11 @@
           />
         </div>
 
-        <!-- New password -->
         <div class="mb-3 position-relative">
           <input
             #newPwd
             type="password"
-            class="form-control"
+            class="form-control input-focus"
             placeholder="New password"
             [type]="showNew ? 'text' : 'password'"
           />
@@ -44,12 +42,11 @@
           />
         </div>
 
-        <!-- Confirm password -->
         <div class="mb-2 position-relative">
           <input
             #confirmPwd
             type="password"
-            class="form-control"
+            class="form-control input-focus"
             placeholder="Confirm new password"
             [type]="showConfirm ? 'text' : 'password'"
           />
@@ -63,7 +60,6 @@
           />
         </div>
 
-        <!-- Error message -->
         <div *ngIf="passwordError" class="text-danger small text-center mt-1 mb-3">
           {{ passwordError }}
         </div>
