+++ frontend/src/app/shared/components/vehicle-form/vehicle-form.css
@@ -9,11 +9,10 @@
   outline: none;
 }
 
-/* Moderni switch stil */
 .custom-switch .form-check-input {
   width: 50px;
   height: 24px;
-  background-color: #ccc;       /* siva boja kad je OFF */
+  background-color: #ccc;  
   border-radius: 12px;
   border: none;
   position: relative;
@@ -27,13 +26,13 @@
   height: 20px;
   top: 2px;
   left: 2px;
-  background-color: #fff;       /* indikator */
+  background-color: #fff;       
   border-radius: 50%;
   transition: transform 0.3s;
 }
 
 .custom-switch .form-check-input:checked {
-  background-color: #333;       /* tamna boja kada je ON */
+  background-color: #333;     
 }
 
 .custom-switch .form-check-input:checked::after {
