+++ frontend/src/app/shared/components/vehicle-form/vehicle-form.spec.ts
